refactor: 파일 구조 변경
Duck 패턴을...따라하려함.
