+++ app/src/main/java/com/example/myapplication/tab2/ImageInfo.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.tab2;
 
 public class ImageInfo {
     private String resId;
