+++ app/src/main/java/com/example/myapplication/tab3/EventDecorator.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.tab3;
 
 import android.app.Activity;
