+++ app/src/main/java/com/example/myapplication/tab3/ToDoDBHelper.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.tab3;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
