+++ app/src/main/res/layout/fragment_1.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".Fragment1">
+    tools:context=".tab1.Fragment1">
 
     <LinearLayout
         android:layout_width="match_parent"
