+++ app/src/main/res/layout/fragment_2.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".Fragment2">
+    tools:context=".tab2.Fragment2">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
