+++ app/src/main/res/layout/fragment_3.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment3">
+    tools:context=".tab3.Fragment3">
 
     <ScrollView
         android:layout_width="fill_parent"
